introduce NftWidgetContent, improve widget performance
