+++ packages/atlas/src/components/_nft/NftWidget/NftWidget.styles.ts
@@ -98,49 +98,3 @@ export const OwnerHandle = styled(Link)`
   justify-self: start;
   text-decoration: none;
 `
-
-export const ButtonGrid = styled.div<SizeProps & { 'data-two-columns'?: boolean }>`
-  display: grid;
-  gap: ${sizes(4)};
-
-  &[data-size=${sizeObj.small}] {
-    gap: ${sizes(2)};
-  }
-
-  &[data-two-columns='true'] {
-    grid-template-columns: 1fr 1fr;
-  }
-`
-
-export const TopBidderTokenContainer = styled.div<SizeProps>`
-  display: flex;
-  align-items: center;
-  position: relative;
-  left: -4px;
-  z-index: 10;
-
-  &::before {
-    display: inline-block;
-    position: absolute;
-    content: '';
-    width: 28px;
-    height: 28px;
-    background: ${cVar('colorBackgroundMuted')};
-    border-radius: 100%;
-    left: -2px;
-    top: -2px;
-  }
-
-  &[data-size=${sizeObj.small}] {
-    &::before {
-      width: 21px;
-      height: 21px;
-      left: -2.5px;
-      top: 1.5px;
-    }
-  }
-`
-
-export const TopBidderContainer = styled.div`
-  display: flex;
-`
